Further group model inheritance tests
