+++ rest_framework/tests/serializer.py
@@ -5,8 +5,6 @@ from rest_framework import serializers
 from rest_framework.tests.models import (HasPositiveIntegerAsChoice, Album, ActionItem, Anchor, BasicModel,
     BlankFieldModel, BlogPost, Book, CallableDefaultValueModel, DefaultValueModel,
     ManyToManyModel, Person, ReadOnlyManyToManyModel, Photo)
-from rest_framework.tests.multitable_inheritance import (ParentModel,
-    ChildModel, AssociatedModel)
 import datetime
 import pickle
 
@@ -96,16 +94,6 @@ class PositiveIntegerAsChoiceSerializer(serializers.ModelSerializer):
 class BrokenModelSerializer(serializers.ModelSerializer):
     class Meta:
         fields = ['some_field']
-
-
-class DerivedModelSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = ChildModel
-
-
-class AssociatedModelSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = AssociatedModel
 
 
 class BasicTests(TestCase):
@@ -1122,39 +1110,3 @@ class DeserializeListTestCase(TestCase):
         self.assertFalse(serializer.is_valid())
         expected = [{}, {'email': ['This field is required.']}, {}]
         self.assertEqual(serializer.errors, expected)
-
-
-class IneritedModelSerializationTests(TestCase):
-
-    def test_multitable_inherited_model_fields_as_expected(self):
-        """
-        Assert that the parent pointer field is not included in the fields
-        serialized fields
-        """
-        child = ChildModel(name1='parent name', name2='child name')
-        serializer = DerivedModelSerializer(child)
-        self.assertEqual(set(serializer.data.keys()),
-                         set(['name1', 'name2', 'id']))
-
-    def test_onetoone_primary_key_model_fields_as_expected(self):
-        """
-        Assert that a model with a onetoone field that is the primary key is
-        not treated like a derived model
-        """
-        parent = ParentModel(name1='parent name')
-        associate = AssociatedModel(name='hello', ref=parent)
-        serializer = AssociatedModelSerializer(associate)
-        self.assertEqual(set(serializer.data.keys()),
-                         set(['name', 'ref']))
-
-    def test_data_is_valid_without_parent_ptr(self):
-        """
-        Assert that the pointer to the parent table is not a required field
-        for input data
-        """
-        data = {
-            'name1': 'parent name',
-            'name2': 'child name',
-        }
-        serializer = DerivedModelSerializer(data=data)
-        self.assertEqual(serializer.is_valid(), True)
